requirements.txt
python-telegram-bot==22.2
supabase==2.16.0
apscheduler==3.11.0
python-dotenv==1.1.1